+++ Runtime/Timeline/CinemachineMixer.cs
@@ -15,6 +15,12 @@ namespace Cinemachine.Timeline
             if (mBrain != null)
                 mBrain.ReleaseCameraOverride(mBrainOverrideId); // clean up
             mBrainOverrideId = -1;
+            DestroyCanvas();
+        }
+
+        public override void PrepareFrame(Playable playable, FrameData info)
+        {
+            mPlaying = info.evaluationType == FrameData.EvaluationType.Playback;
         }
 
         public override void ProcessFrame(Playable playable, FrameData info, object playerData)
@@ -37,48 +43,126 @@ namespace Cinemachine.Timeline
             int activeInputs = 0;
             ICinemachineCamera camA = null;
             ICinemachineCamera camB = null;
-            float camWeight = 1f;
+            Texture texA = null;
+            Texture texB = null;
+            float camWeightB = 1f;
+            float texWeightB = 0;
             for (int i = 0; i < playable.GetInputCount(); ++i)
             {
                 CinemachineShotPlayable shot
                     = ((ScriptPlayable<CinemachineShotPlayable>)playable.GetInput(i)).GetBehaviour();
                 float weight = playable.GetInputWeight(i);
-                if (shot != null && shot.VirtualCamera != null
+                if (shot != null && shot.IsValid
                     && playable.GetPlayState() == PlayState.Playing
                     && weight > 0.0001f)
                 {
-                    if (activeInputs == 1)
-                        camB = camA;
-                    camWeight = weight;
-                    camA = shot.VirtualCamera;
-                    ++activeInputs;
-                    if (activeInputs == 2)
+                    if (shot.m_StoryboardImage)
+                    {
+                        texA = texB;
+                        texWeightB = weight;
+                        texB = shot.m_Image;
+                    }
+                    if (!shot.m_StoryboardImage || shot.VirtualCamera != null)
+                    {
+                        camA = camB;
+                        camWeightB = weight;
+                        camB = shot.VirtualCamera;
+                    }
+                    if (++activeInputs == 2)
                         break;
                 }
             }
 
-            float deltaTime = info.deltaTime;
+            // Override the Cinemachine brain with our results
+            mBrainOverrideId = mBrain.SetCameraOverride(
+                    mBrainOverrideId, camA, camB, camWeightB, GetDeltaTime(info.deltaTime));
+
+            // Display the storyboard overlays
+            DisplayCanvasOverlays(texA, texB, texWeightB);
+        }
+
+        float mLastOverrideFrame;
+        float GetDeltaTime(float deltaTime)
+        {
             if (!mPlaying)
             {
                 if (mBrainOverrideId < 0)
                     mLastOverrideFrame = -1;
                 float time = Time.realtimeSinceStartup;
                 deltaTime = Time.unscaledDeltaTime;
-                if (!Application.isPlaying && (mLastOverrideFrame < 0 || time - mLastOverrideFrame > Time.maximumDeltaTime))
+                if (!Application.isPlaying 
+                    && (mLastOverrideFrame < 0 || time - mLastOverrideFrame > Time.maximumDeltaTime))
+                {
                     deltaTime = -1;
+                }
                 mLastOverrideFrame = time;
             }
-
-            // Override the Cinemachine brain with our results
-            mBrainOverrideId = mBrain.SetCameraOverride(
-                    mBrainOverrideId, camB, camA, camWeight, deltaTime);
-
+            return deltaTime;
         }
-        float mLastOverrideFrame;
 
-        public override void PrepareFrame(Playable playable, FrameData info)
+
+        // Storyboard support
+        GameObject mCanvas;
+        UnityEngine.UI.RawImage mRawImageA;
+        UnityEngine.UI.RawImage mRawImageB;
+
+        void CreateCanvas()
         {
-            mPlaying = info.evaluationType == FrameData.EvaluationType.Playback;
+            mCanvas = new GameObject("_CM_Mixer_canvas", typeof(RectTransform));
+            mCanvas.hideFlags = HideFlags.HideAndDontSave;
+
+            var c = mCanvas.AddComponent<Canvas>();
+            c.renderMode = RenderMode.ScreenSpaceOverlay;
+
+            var go = new GameObject("RawImageA", typeof(RectTransform));
+            go.transform.SetParent(mCanvas.transform);
+            mRawImageA = go.AddComponent<UnityEngine.UI.RawImage>();
+
+            go = new GameObject("RawImageB", typeof(RectTransform));
+            go.transform.SetParent(mCanvas.transform);
+            mRawImageB = go.AddComponent<UnityEngine.UI.RawImage>();
+        }
+        
+        void DestroyCanvas()
+        {
+            if (mCanvas != null)
+                Object.DestroyImmediate(mCanvas);
+            mCanvas = null;
+            mRawImageA = null;
+            mRawImageB = null;
+        }
+
+        void DisplayCanvasOverlays(Texture texA, Texture texB, float texWeightB)
+        {
+            if (texB == null)
+            {
+                if (mCanvas != null)
+                    mCanvas.SetActive(false);
+            }
+            else
+            {
+                if (mCanvas == null)
+                    CreateCanvas();
+                mCanvas.SetActive(true);
+
+                Color color = Color.white;
+                mRawImageA.texture = texA;
+                mRawImageA.color = color;
+                mRawImageA.rectTransform.localPosition = Vector3.zero;
+                mRawImageA.rectTransform.localRotation = Quaternion.identity;
+                mRawImageA.rectTransform.localScale = Vector3.one;
+                mRawImageA.rectTransform.sizeDelta = new Vector2(Screen.width, Screen.height);
+                mRawImageA.gameObject.SetActive(texA != null);
+
+                mRawImageB.texture = texB;
+                color.a = texWeightB;
+                mRawImageB.color = color;
+                mRawImageB.rectTransform.localPosition = Vector3.zero;
+                mRawImageB.rectTransform.localRotation = Quaternion.identity;
+                mRawImageB.rectTransform.localScale = Vector3.one;
+                mRawImageB.rectTransform.sizeDelta = new Vector2(Screen.width, Screen.height);
+                mRawImageB.gameObject.SetActive(texB != null);
+            }
         }
     }
 }