+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.1.11] - 2018-01-01
+## [2.1.11] - 2018-01-21
 ### Bugfixes and minor enhancements.  Also some restructuring in prep for UPM
 - Project restructure: Removed Base, Timeline, and PostFX folders from project root
+- New OverlayImage extension to put screen-space overlays on vcams.  Great for storyboard placeholders.
 - Added support for custom blend curves
 - Lookahead: added Ignore Y Axis Movement option
 - Added support for cascading blends (i.e. blending from mid-blend looks better)
@@ -14,11 +15,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - FreeLook: added Y Axis recentering
 - Path: Added Normalized Path units option: 0 is start of path, 1 is end.
 - Path: added length display in inspector
+- Timeline Clip Editor: vcam sections are now collapsible
 - Bugfix: OrbitalTransposer works better with WorldUp overrides
 - Bugfix: Remove StateDrivenCamera "not playing a controller" warning
 - Bugfix: DepthOfField blending was wrong in PPv2
 - Bugfix: Handle exceptions thrown by assemblies that don't want to be introspected
 - Bugfix: vcams following physics objects snapped to origin after exiting play mode
+- Bugfix: predictor now supports time pause
+- Bugfix: Moved StartCoroutine in Brain to OnEnable()
 
 ## [2.1.10] - 2017-11-28
 ### This is the first UPM release of *Unity Package Cinemachine*.